wc_port.c: NETOS: Fix typo.
This 1 line patch fixes a typo (or copy/paste error) in wc_port.c
contained under the NETOS section.

TESTSUITE_THREAD_STACK_SZ is currently used in tx_thread_create()
call for the size of the stack.

However, WOLFSSL_NETOS_STACK_SZ should be used instead.

The WOLFSSL_NETOS_STACK_SZ is used a few lines above to actually allocate
the stack and its size, so the same value should be used to
tell tx_thread_create() what the size of the stack is.